frontend fix
Fix bug + checkout interface

checkout

+++ app/Http/Controllers/PagesController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Cart;
+use App\Models\Order;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
@@ -11,14 +12,17 @@ class PagesController extends Controller
 {
     public function showCart()
     {
+        $user = Auth::user();
         $products = DB::table('products')->join('cart', 'products.id', '=', 'cart.product_id')->distinct()->get();
+        $user->cart = $products->count();
+        $user->save();
 
         return view('sites.cart', compact('products'));
     }
 
     public function addToCart(Request $request)
     {
-        $cart = Cart::where('product_id', '=', $request->product_id)->first();
+        $cart = Cart::where('product_id', '=', $request->id)->first();
         if ($cart != null) {
             $cart->quantity += $request->quantity;
             $cart->save();
@@ -27,19 +31,76 @@ class PagesController extends Controller
         else {
             $cart = new Cart();
             $cart->user_id = Auth::user()->id;
-            $cart->product_id = $request->product_id;
+            $cart->product_id = $request->id;
             $cart->quantity = $request->quantity;
             $cart->save();
+            $user = Auth::user();
+            $user->cart += 1;
+            $user->save();
         }
 
-        return redirect(route('showCart', Auth::user()->id));
+        return redirect(route('showCart'));
     }
 
     public function deleteCart(Request $request)
     {
-        $cart = Cart::where('product_id', '=', $request->id)->first();
-        Cart::destroy($cart->id);
+        Cart::destroy($request->id);
+        $user = Auth::user();
+        $user->cart -= 1;
+        $user->save();
 
-        return redirect(route('showCart', Auth::user()->id));
+        return redirect(route('showCart'));
+    }
+
+    public function checkout()
+    {
+        $user = Auth::user();
+        $products = DB::table('products')->join('cart', 'products.id', '=', 'cart.product_id')->distinct()->get();
+
+        return view('sites.checkout', compact('user', 'products'));
+    }
+
+    public function saveCart(Request $request)
+    {
+        $i = 0;
+        $carts = Cart::orderBy('created_at', 'asc')->get();
+        $count = Auth::user()->cart;
+        foreach ($carts as $cart) {
+            $cart->quantity = $_POST["qty"][ $i++ ];
+            if ($cart->quantity == 0) {
+                Cart::destroy($cart->id);
+                $user = Auth::user();
+                $user->cart -= 1;
+                $user->save();
+            }
+            else {
+                $cart->save();
+            }
+        }
+
+        return redirect(route('checkout'));
+    }
+
+    public function thankyou()
+    {
+        return view('sites.thankyou');
+    }
+
+    public function checkoutSubmit(Request $request)
+    {
+        $user = Auth::user();
+        $order = new Order();
+        $order->user_id = $user->id;
+        $order->product_count = $user->cart;
+        $order->sum = $request->sum;
+        $order->save();
+        $user->cart = 0;
+        $user->save();
+        $carts = Cart::where('user_id', '=', $user->id)->get();
+        foreach ($carts as $cart) {
+            Cart::destroy($cart->id);
+        }
+
+        return redirect(route('thankyou'));
     }
 }
